hub_component/requirements.txt
asgiref==3.8.1
Django==5.1.4
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
djangorestframework>=3.15.2
watchgod
celery>=5.3.4
redis>=5.0.0
docker
django-cors-headers
pytest
pytest-django
factory_boy
pytest-mock
freezegun
pytest-cov
matplotlib
networkx